backend/requirements.txt
pillow 
ipfshttpclient
flask
web3
python-dotenv
gunicorn 
gevent
Flask-Limiter
Flask-Limiter[redis]
psutil
redis